"add warm sms"

+++ app/src/main/java/learn/android/kangel/mycontacts/fragments/SearchFragment.java
@@ -132,7 +132,7 @@ public class SearchFragment extends android.support.v4.app.ListFragment implemen
         if (c != null) {
             c.moveToPosition(position);
             String lookUpKey = c.getString(7);
-            int contactId = c.getInt(6);
+            long contactId = c.getLong(6);
             Intent intent = new Intent(getActivity(), ContactDetailActivity.class);
             intent.putExtra("lookUpKey", lookUpKey);
             intent.putExtra("contactId", contactId);
